+++ cropper/src/main/java/com/canhub/cropper/CropException.kt
@@ -3,8 +3,15 @@ package com.canhub.cropper
 import android.net.Uri
 
 sealed class CropException(message: String) : Exception(message) {
-    class CancellationException : CropException("cropping has been cancelled by the user")
-    class FailedToLoadBitmapException(uri: Uri, message: String?) :
-        CropException("Failed to load sampled bitmap: $uri\r\n$message")
-    class FailedToDecodeImageException(uri: Uri) : CropException("Failed to decode image: $uri")
+    class Cancellation : CropException("$EXCEPTION_PREFIX cropping has been cancelled by the user")
+    class FailedToLoadBitmap(uri: Uri, message: String?) :
+        CropException("$EXCEPTION_PREFIX Failed to load sampled bitmap: $uri\r\n$message")
+
+    class FailedToDecodeImage(uri: Uri) :
+        CropException("$EXCEPTION_PREFIX Failed to decode image: $uri")
+
+    companion object {
+
+        const val EXCEPTION_PREFIX = "crop:"
+    }
 }